BZ-1310739 - Performance degradation with REST API on WebSphere: JAXBContext CDI use
