+++ kie-remote/kie-remote-services/src/main/java/org/kie/remote/services/rest/jaxb/DynamicJaxbContextFilter.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
@@ -27,6 +27,7 @@ import javax.servlet.ServletException;
 import javax.servlet.ServletRequest;
 import javax.servlet.ServletResponse;
 import javax.servlet.http.HttpServletRequest;
+import javax.xml.bind.JAXBContext;
 
 import org.jbpm.kie.services.api.DeploymentIdResolver;
 import org.kie.remote.services.cdi.DeploymentInfoBean;
@@ -34,17 +35,25 @@ import org.kie.services.client.serialization.JaxbSerializationProvider;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+/**
+ * This is a {@link Filter} implemenation that's responsible for 2 things:<ol>
+ * <li>Resolving the deploymentId value in the HTTP (REST) request so that we can</li>
+ * <li>retrieve (and if necessary create) the {@link JAXBContext} for the given deployment</li>
+ * </ol>
+ * Lastly, once that {@link JAXBContext} has been created, it's set as a {@link ThreadLocal} instance, since
+ * the request (particularly, the de/serialization of the request/response) will all happen in this thread.
+ */
 public class DynamicJaxbContextFilter implements Filter {
 
     private static final Logger logger = LoggerFactory.getLogger(DynamicJaxbContextFilter.class);
     // private FilterConfig _filterConfig; // not used
-  
-    // "**" not accepted in URL's.. 
+
+    // "**" not accepted in URL's..
     public static final String DEFAULT_JAXB_CONTEXT_ID = "**DEFAULT";
 
     @Inject
     private DeploymentInfoBean deploymentInfoBean;
-    
+
     public void init(FilterConfig filterConfig) throws ServletException {
         // do nothing
     }
@@ -59,50 +68,50 @@ public class DynamicJaxbContextFilter implements Filter {
 
         DynamicJaxbContext.setDeploymentJaxbContext(deploymentId);
         logger.debug("JAXBContext retrieved and set for for '{}'", deploymentId);
-        try { 
+        try {
             chain.doFilter(request, response);
-        } finally { 
+        } finally {
             DynamicJaxbContext.clearDeploymentJaxbContext();
         }
     }
 
     @Override
     public void destroy() {
-        // _filterConfig = null; // not used
+        // not used
     }
 
     // pkg static for tests
     static String getDeploymentId(HttpServletRequest request) {
         String deploymentId = null;
-    
+
         // extract from header
-        deploymentId = request.getHeader(JaxbSerializationProvider.EXECUTE_DEPLOYMENT_ID_HEADER); 
-        if( ! emptyDeploymentId(deploymentId) ) { 
-           return deploymentId; 
+        deploymentId = request.getHeader(JaxbSerializationProvider.EXECUTE_DEPLOYMENT_ID_HEADER);
+        if( ! emptyDeploymentId(deploymentId) ) {
+           return deploymentId;
         }
-        
+
         // extract from the proper url
         String requestUri = request.getRequestURI();
         String [] urlParts = requestUri.split("/");
-        for( int i = 0; i < urlParts.length; ++i ) { 
+        for( int i = 0; i < urlParts.length; ++i ) {
            if( urlParts[i].equals("deployment")
-               || urlParts[i].equals("runtime") ) { 
-              if( i+1 < urlParts.length ) { 
+               || urlParts[i].equals("runtime") ) {
+              if( i+1 < urlParts.length ) {
                   deploymentId = urlParts[i+1];
                   if( ! emptyDeploymentId(deploymentId) ) {
-                     return deploymentId; 
+                     return deploymentId;
                   }
                   break;
               }
            }
         }
-        
+
         // get parameter
         deploymentId = request.getParameter("deploymentId");
-        if( ! emptyDeploymentId(deploymentId) ) { 
-           return deploymentId; 
+        if( ! emptyDeploymentId(deploymentId) ) {
+           return deploymentId;
         }
-        
-        return DEFAULT_JAXB_CONTEXT_ID; 
-    } 
+
+        return DEFAULT_JAXB_CONTEXT_ID;
+    }
 }
